Add support for undefined_tramp and closure_tramp in read-only space.
  * These are the two "tricky" trampolines, in that genesis needs
to know where they are in order to cold-load fasls, and where
undefined_tramp in particular is in order to even prepare to
cold-load fasls.

  * Add a concept of a "preload file" to genesis, to be loaded
after creating NIL but before creating any FDEFN objects.

  * Use a target feature to switch genesis to using an assembler
routine instead of a foreign symbol for the trampolines.

  * And use our new :NOT-GENESIS build flag to set up a file that
will contain the new trampolines, and arrange for it to be loaded
as the preload file, again under the same target feature.

  * As far as it goes, all this works, but it almost certainly has
some rough edges to it.  It probably isn't quite ready for
upstream yet.